[Player]

+++ Game/Game/core/Box2D.cpp
@@ -1,6 +1,6 @@
 #include "Box2D.h"
 
-void Box2D::setPos(int x, int y)
+void Box2D::setPos(float x, float y)
 {
 	aabb.min.x = x;
 	aabb.min.y = y;
@@ -8,7 +8,7 @@ void Box2D::setPos(int x, int y)
 	aabb.max.y = y + h;
 }
 
-void Box2D::moveBy(int xRel, int yRel)
+void Box2D::moveBy(float xRel, float yRel)
 {
 	aabb.min.x += xRel;
 	aabb.min.y += yRel;
